pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Igorik <egorsborsikov250575@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
supabase = "^2.13.0"
fastapi = "^0.115.11"
uuid = "^1.30"
asyncio = "^3.4.3"
uvicorn = "^0.34.0"
pydantic-settings = "^2.8.1"
python-dotenv = "^1.1.0"
python-telegram-bot = "^22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
